ai-test/contest_trade_refactor/requirements.txt
tqdm
requests
matplotlib
tushare
loguru
openai
tiktoken
finnhub-python
typer
pydantic
questionary
prompt_toolkit<3.0.30
PyYAML
pandas
langgraph
rich
tabulate
akshare
crawl4ai
lightgbm